+++ apps/web/client/src/app/layout.tsx
@@ -1,7 +1,7 @@
 import '@/styles/globals.css';
 import '@onlook/ui/globals.css';
 
-import { Modals } from '@/components/modal';
+import { Modals } from '@/components/ui/modal';
 import { TRPCReactProvider } from '@/trpc/react';
 import { type Metadata } from 'next';
 import { NextIntlClientProvider } from 'next-intl';
