+++ apps/web/client/src/components/ui/modal/settings/project/reinstall-button.tsx
