+++ apps/web/client/src/components/ui/modal/settings/site/favicon.tsx
